% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{deriv.polynR}
\alias{deriv.polynR}
\title{Derivative of a polynom}
\usage{
\method{deriv}{polynR}(p)
}
\arguments{
\item{p}{an integer vector of  coefficients}
}
\value{
the algebric form of the derivative of a polynom
}
\description{
Derivative of a polynom
}
\examples{
print.polynR(deriv.polynR(c(6,2,10,2,32)))
}
