% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{derive.plynR}
\alias{derive.plynR}
\title{Derivative polynom}
\usage{
\method{derive}{plynR}(p)
}
\arguments{
\item{p}{an integer vector of  coefficients}
}
\value{
the algebric form of the derivative of a polynom
}
\description{
This function allows the user to compute the derivative of a polynom from class polynR
}
\examples{
\dontrun{printPol.plynR(derive.plynR(c(1,1,10)))}
}
