% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{dft1_iter}
\alias{dft1_iter}
\title{Iterative Form for Fast Fourier Transform}
\usage{
dft1_iter(x, k)
}
\arguments{
\item{x}{a complex vector}

\item{k}{an integer}
}
\value{
Imaginary part of fast fourier transform
}
\description{
Iterative Form for Fast Fourier Transform
}
\examples{

}
