% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{dft1_matrix}
\alias{dft1_matrix}
\title{Matrix form of Fast Fourier Transform}
\usage{
dft1_matrix(x,k)
}
\arguments{
\item{x}{a complex vector}

\item{k}{an integer}
}
\value{
Imaginary part of fast fourier inverse
}
\description{
This function allows the user to compute the fast fourier transform with a matrix method.
}
\examples{
\dontrun{vec_com<-complex(real = sample(10), imaginary = sample(10))
dft1_matrix(vec_com,3)}
}
