% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{dft1_naive}
\alias{dft1_naive}
\title{Naive form Fast Fourier Inverse}
\usage{
dft1_naive(x, k)
}
\arguments{
\item{x}{a complex vector}

\item{k}{an integer}
}
\value{
Imaginary part of fast fourier inverse
}
\description{
Naive form Fast Fourier Inverse
}
