% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{dft_factory}
\alias{dft_factory}
\title{Customized Fast Fourier method computation}
\usage{
dft_factory(xk,FunF)
}
\arguments{
\item{xk}{a complex vector}

\item{FunF}{method to be used}
}
\value{
Fast Fourier Transform according to the method specified
}
\description{
Run a pre-specified method by user.
}
\examples{
\dontrun{vec_com<-complex(real = sample(10), imaginary = sample(10))
dft_factory(vec_com,3)}
}
