% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{discrete_inv}
\alias{discrete_inv}
\title{Inverse Discrete Distribution}
\usage{
discrete_inv(vec)
}
\arguments{
\item{vec}{an integer vector}
}
\value{
the inverse of discrete distribution
}
\description{
Inverse Discrete Distribution
}
\examples{
vec<-c(2,10,4,6)
discrete_inv(vec)
}
