% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{discrete_inverse_tran}
\alias{discrete_inverse_tran}
\title{Discrete Inverse Transformation}
\usage{
discrete_inverse_tran(sim, trans)
}
\arguments{
\item{sim}{a method of generator random variable}

\item{trans}{a customized transformation function}
}
\value{
a vector of a transformed realisation
}
\description{
Discrete Inverse Transformation
}
\examples{
vec<-c(2,10,4,6)
discrete_inverse_tran(simulateur1(100), vec**2)
}
