% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{fft_ct2}
\alias{fft_ct2}
\title{Scalable Fast Fourier Transform}
\usage{
fft_ct2(x)
}
\arguments{
\item{x}{an integer}
}
\value{
Fast Fourier Transform sequence
}
\description{
Scalable Fast Fourier Transform
}
