% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{fib_mulPos_rec}
\alias{fib_mulPos_rec}
\title{Fibonacci Serie}
\usage{
fib_mulPos_rec(k)
}
\arguments{
\item{k}{an integer}
}
\value{
a fibonacci sequence
}
\description{
This function allows to compute the Fibonacci serie
}
\examples{
\dontrun{fib_mulPos_sec(3)}
}
