% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{fois.plynR}
\alias{fois.plynR}
\title{Fois polynoms}
\usage{
\method{fois}{plynR}(p1, p2)
}
\arguments{
\item{p1}{an integer coefficient (first polynom)}

\item{p2}{an integer coefficient (first polynom)}
}
\value{
multiplication of two polynoms from the input coefficients
}
\description{
Multiply two polynomias from class polynR
}
\examples{
\dontrun{printPol.plynR(fois.plynR(c(1,1,10), c(1,2,10)))}
}
