% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{fois.polynR}
\alias{fois.polynR}
\title{Multiply two polynomial series}
\usage{
\method{fois}{polynR}(...)
}
\arguments{
\item{serie1}{first polynom}

\item{serie2}{second polynom}
}
\value{
multiplication of two polynoms
}
\description{
Multiply two polynomial series
}
\examples{
print.polynR(fois.polynR(c(1,1,10), c(1,2,10)))
}
