% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{NbSerie}
\alias{NbSerie}
\title{Sequence of a bounded arithmetic operation}
\usage{
NbSerie(l)
}
\arguments{
\item{l}{an integer vector}
}
\value{
the longest vector of realisation inside the bound.
}
\description{
Sequence of a bounded arithmetic operation
}
\examples{
NbSerie(c(2,2,5,6,6,6,7,9))
}
