% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{plus.plynR}
\alias{plus.plynR}
\title{Plus polynoms}
\usage{
\method{plus}{plynR}(p1, p2)
}
\arguments{
\item{p1}{an integer coefficient (first polynom)}

\item{p2}{an integer coefficient (second polynom)}
}
\value{
sum of two polynoms from the input coefficients
}
\description{
Add two polynoms from class polynR
}
\examples{
\dontrun{printPol.plynR(plus.polynR(c(1,1,10), c(1,2,10)))}
}
