% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{plus.polynR}
\alias{plus.polynR}
\title{Add two polynomial series}
\usage{
\method{plus}{polynR}(...)
}
\arguments{
\item{serie1}{first polynom}

\item{serie2}{second polynom}
}
\value{
sum of two polynoms
}
\description{
Add two polynomial series
}
\examples{
plus.polynR(c(1,1,10), c(1,2,10))
print.polynR(plus.polynR(c(1,1,10), c(1,2,10)))
}
