% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{plynR}
\alias{plynR}
\title{Class Polynomial}
\usage{
plynR(coefs)
}
\arguments{
\item{coefs}{an integer vector of  coefficients}
}
\value{
a polynomial class
}
\description{
This class allows to construct a family of polynomials in order to manipulate arithmetic operations on polynomials.
eg. sum, multiplication, roots identification.
}
\examples{
\dontrun{plynR(c(1,2,-3))}
}
