% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{polynR}
\alias{polynR}
\title{Polynomial}
\usage{
polynR(coefs)
}
\arguments{
\item{coefs}{an integer vector of  coefficients}
}
\value{
a polynomial class
}
\description{
Polynomial
}
\examples{
polynR(c(1,2,-3))
}
