% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{print.polynR}
\alias{print.polynR}
\title{Print Polynomials
A generic method of class Polynom.}
\usage{
\method{print}{polynR}(p)
}
\arguments{
\item{p}{an integer vector of  coefficients}
}
\value{
the algebric form of a polynom.
}
\description{
Print Polynomials
A generic method of class Polynom.
}
\examples{
print.polynR(c(1,2,30))
}
