% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{printPol.plynR}
\alias{printPol.plynR}
\title{Print Polynomials}
\usage{
\method{printPol}{plynR}(p)
}
\arguments{
\item{p}{an integer vector of  coefficients}
}
\value{
print the algebric form of a class polynom.
}
\description{
A generic method of class Polynom that allows the user to print the algebric form of a polynom from a given input coefficients.
}
\examples{
\dontrun{printPol.plynR(c(1,2,3))}
}
