% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{racines.plynR}
\alias{racines.plynR}
\title{Racines polynoms}
\usage{
\method{racines}{plynR}(p)
}
\arguments{
\item{p}{an integer vector of  coefficients}
}
\value{
return reals and imaginary solutions of a polynom.
}
\description{
This function allows the user to determine the roots of a polynom from class polynR
}
\examples{
\dontrun{printPol(racines.plynR(c(3,-4,5)))}
}
