% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{racines.polynR}
\alias{racines.polynR}
\title{Roots of a polynom}
\usage{
\method{racines}{polynR}(p)
}
\arguments{
\item{p}{an integer vector of  coefficients}
}
\value{
Reals and imaginaries solutions of a polynom.
}
\description{
Roots of a polynom
}
\examples{
print.polynR(Re(racines.polynR(polynR(c(3,-4,5)))))
}
