% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_v2.R
\name{realization_sum}
\alias{realization_sum}
\title{Generator of a sum of random variables}
\usage{
realization_sum(vec1,vec2,trans=identity)
}
\arguments{
\item{vec1}{an integer vector of realization (first)}

\item{vec2}{an integer vector of realization (second)}

\item{trans}{a transformation function}
}
\value{
a vector of a transformed sum from the input vectors
}
\description{
This function allows to the user to generate the sum of two generators.
The user could specify a transformation for the input otherwise it will return the identity transformation.
}
\examples{
\dontrun{
simulator_sum<-function(v1,v2,trans=identity){
lapply(list(length(c(v1,v2))),realization_sum(v1,v2, trans))};
simulator_sum(c(1,2,3),c(2,3,4), identity);
f2<-function(x) x**2;
sim_sum<-simulator_sum(c(1,2,3),c(2,3,4), f2);
sim_sum
}
}
