% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_v2.R
\name{sequence_long}
\alias{sequence_long}
\title{Sequence Generator}
\usage{
sequence_long(sets = c(2L, 2L, 5L, 6L, 6L, 6L, 7L, 9L),
lowerb = -1e10,
upperb = 1e10,
nb_iters = 1e5L)
}
\arguments{
\item{sets}{an integer vector}

\item{lowerb}{lower bound}

\item{upperb}{upper bound}

\item{nb_iters}{number of iterations}
}
\value{
the longest sequence inside the given bound
}
\description{
This function allows to calculate the longest bounded sequence given a defined arithmetic operation.
}
\examples{
\dontrun{
sequence_long(sets = c(2L, 2L, 5L, 6L, 6L, 6L, 7L, 9L),
lowerb = -1e10,
upperb = 1e10,
seq_length=5000)
}
}
