% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulateur1}
\alias{simulateur1}
\title{Generator of Discrete distribution}
\usage{
simulateur1(n)
}
\arguments{
\item{n}{an integer vector}
}
\value{
a sequence of discrete random variable
}
\description{
Generator of Discrete distribution
}
\examples{
simulateur1(10)
}
