% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulateur2}
\alias{simulateur2}
\title{Generator of a Transformed Discrete distribution}
\usage{
simulateur2(sim, trans)
}
\arguments{
\item{sim}{a method of generator random variable}

\item{trans}{a customized transformation function}
}
\value{
a vector of a transformed realisation
}
\description{
Generator of a Transformed Discrete distribution
}
\examples{
 
vec<-c(2,10,4,6)
simulateur2(simulateur1(100), vec**2).
}
