% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulateur_bar}
\alias{simulateur_bar}
\title{Histogram of distributed realizations}
\usage{
simulateur_bar(sim)
}
\arguments{
\item{sim}{a generator of random variable}
}
\value{
histogram of the distribution returned
}
\description{
Histogram of distributed realizations
}
\examples{
n<-100
simulateur_bar(simulateur1(n))
simulateur_bar(simulateur2(simulateur1(n), vec**2))
}
