% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulateur_prop}
\alias{simulateur_prop}
\title{Generator of proportion}
\usage{
simulateur_prop(sim)
}
\arguments{
\item{sim}{a vector of simulated distribution}
}
\value{
a vector of simulated distribution
}
\description{
Generator of proportion
}
\examples{
simulateur_prop(simulateur1(100))
simulateur_prop(simulateur2(simulateur1(100), vec**2))
simulateur_prop(simulateur_sum(simulateur1(n),simulateur2(simulateur1(n), vec**2)))
}
