% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulateur_sum}
\alias{simulateur_sum}
\title{Sum of two generators}
\usage{
simulateur_sum(sim1, sim2)
}
\arguments{
\item{sim1}{a vector of simulated distribution}

\item{sim2}{a vector of simulated distribution}
}
\value{
a vector of simulated distribution
}
\description{
Sum of two generators
}
\examples{
simulateur_sum(simulateur1(10),simulateur1(50))
simulateur_sum(simulateur1(100),simulateur2(simulateur1(100), vec**2))
simulateur_sum(simulateur1(10),simulateur1(50))
}
