% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_v2.R
\name{simulator}
\alias{simulator}
\title{Generator of a transformed Discrete distribution}
\usage{
simulator(x_vec, trans = identity)
}
\arguments{
\item{x_vec}{an integer vector}

\item{trans}{a transformation function}
}
\value{
a sequence of discrete random variable from the input distribution
}
\description{
This function generate a vector of realization from the same distribution of the input vector.
The user can specify a transformation on the input vector otherwise it will return the identity transformation.
}
\examples{
\dontrun{
f2<-function(x) x**2;
xs<-c(10,3,5,20,4,6);
simulator_tran<-function(xs,trans=identity) lapply(list(length(xs)),simulator(xs, trans))
sim_tran<-simulator_tran(xs)
}
}
