% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_v2.R
\name{simulator_bar}
\alias{simulator_bar}
\title{Histogram of distributed realizations}
\usage{
simulator_bar(sim)
}
\arguments{
\item{sim}{a simulator}
}
\value{
a histogram of the realizations
}
\description{
This function allows to the user to visualize the empirical distribution of the realized input vector
The user could specify a transformation of the input vector otherwise it will return he identity transformation.
}
\examples{
\dontrun{
f2<-function(x) x**2;
xs<-c(10,3,5,20,4,6);
sim_tran<-simulator_tran(xs)
sim_tran2<-simulator_tran(xs,f2);
simulator_prop(sim_tran);
simulator_bar(sim_tran);
simulator_bar(sim_tran2);
}
}
