% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{summary.polynR}
\alias{summary.polynR}
\title{Summary of a polynom}
\usage{
\method{summary}{polynR}(object)
}
\arguments{
\item{object}{a polynom object}
}
\value{
summary of a polynom.
}
\description{
Summary of a polynom
}
\examples{
summary.polynR(polynR(c(3,-4,5)))
}
