% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomials.R
\name{summaryPol.plynR}
\alias{summaryPol.plynR}
\title{Summary polynom}
\usage{
summaryPol.plynR(object)
}
\arguments{
\item{object}{a polynom object from class ploynR}
}
\value{
summary of a polynom.
}
\description{
This function allows the user to return the summary of a polynom from class polynR and gives its factorization
}
\examples{
\dontrun{summaryPol.plynR(plynR(c(1,2,3)))}
}
