% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_v2.R
\name{tracking_operator}
\alias{tracking_operator}
\title{Sequence Tracker}
\usage{
tracking_operator(invec,k)
}
\arguments{
\item{invec}{a list of computed operations}

\item{k}{the kth result from the list (correspond to the length of the sequence 1e5L by default)}
}
\value{
the decomposition of the given result
}
\description{
This function allows to track the decomposition of a selected operation.
}
\examples{
\dontrun{
w<-sequence_long(sets = c(2L, 2L, 5L, 6L, 6L, 6L, 7L, 9L),
lowerb = -1e10,
upperb = 1e10,
nb_iters = 1e5L)
tracking_operator(w,k=10,4)
tracking_operator(w,k=100,5)
}
}
