% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{dft1_matrix}
\alias{dft1_matrix}
\title{Matrix form of Fast Fourier Transform}
\usage{
dft1_matrix(x, k)
}
\arguments{
\item{x}{a complex vector}

\item{k}{an integer}
}
\value{
Imaginary part of fast fourier inverse
}
\description{
Matrix form of Fast Fourier Transform
}
