% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{dft1_naive}
\alias{dft1_naive}
\title{Naive Fast Fourier Inverse}
\usage{
dft1_naive(x,k)
}
\arguments{
\item{x}{a complex vector}

\item{k}{an integer}
}
\value{
Imaginary part of fast fourier inverse
}
\description{
This function allows the user to compute the fast fourier transform with a naive method.
}
\examples{
\dontrun{vec_com<-complex(real = sample(10), imaginary = sample(10))
dft1_naive(vec_com,3)}
}
