% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_v2.R
\name{tracking_operator}
\alias{tracking_operator}
\title{Sequence Tracker}
\usage{
tracking_operator(data,k,i)
}
\arguments{
\item{k}{the k_th result from the list (k correspond to the length of the sequence 5000 by default)}

\item{i}{the i_th operation performed randomly (5 operations performed for each set of number)}

\item{data}{a list of stored operations}
}
\value{
the decomposition of the given result
}
\description{
This function allows to track the decomposition of a selected operation.
}
\examples{
\dontrun{
w<-sequence_long(sets = c(2L, 2L, 5L, 6L, 6L, 6L, 7L, 9L),
lowerb = -1e10,
upperb = 1e10, seq_length= 5000)
tracking_operator(sequence_long = w, k=30,i=4);
tracking_operator(sequence_long = w, k=10, i=5)
}
}
